Cleanup Terminal from deprecated code and extension interfaces
As we now have a new namespace we can use the opportunity to remove any
deprecated code and squash previous extension interfaces into one.
