+++ .github/workflows/doCleanCode.yml
@@ -14,6 +14,6 @@ jobs:
       author: Eclipse Platform Bot <platform-bot@eclipse.org>
       do-quickfix: false
       do-cleanups: true
-      bundle-folders: ant/*/ debug/*/ resources/bundles/*/ runtime/bundles/*/ team/bundles/*/ ua/*/ update/*/ platform/*/
+      bundle-folders: ant/*/ debug/*/ resources/bundles/*/ runtime/bundles/*/ team/bundles/*/ ua/*/ update/*/ platform/*/ terminal/plugins/*/
     secrets:
       token: ${{ secrets.PLATFORM_BOT_PAT }}
