--- terminal/plugins/org.eclipse.tm.terminal.control/src/org/eclipse/terminal/internal/control/ITerminalListener2.java
@@ -1,29 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2015, 2018 Wind River Systems, Inc. and others. All rights reserved.
- * This program and the accompanying materials are made available under the terms
- * of the Eclipse Public License 2.0 which accompanies this distribution, and is
- * available at https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- * Wind River Systems - initial API and implementation
- *******************************************************************************/
-package org.eclipse.terminal.internal.control;
-
-/**
- * Terminal listener allowing to listen to terminal selection changes.
- *
- * @since 4.1
- */
-public interface ITerminalListener2 extends ITerminalListener {
-
-	/**
-	 * selection has been changed internally e.g. select all
-	 * clients might want to react on that
-	 * NOTE: this does not include mouse selections
-	 * those are handled in separate MouseListeners
-	 * TODO should be unified
-	 */
-	void setTerminalSelectionChanged();
-}
