+++ terminal/plugins/org.eclipse.tm.terminal.control/src/org/eclipse/terminal/internal/provisional/api/ITerminalControl.java
@@ -17,12 +17,11 @@
 package org.eclipse.terminal.internal.provisional.api;
 
 import java.io.OutputStream;
-import java.io.UnsupportedEncodingException;
 import java.nio.charset.Charset;
 
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Shell;
-import org.eclipse.terminal.internal.control.ITerminalListener3.TerminalTitleRequestor;
+import org.eclipse.terminal.internal.control.ITerminalListener.TerminalTitleRequestor;
 
 /**
  * Represents the terminal view as seen by a terminal connection.
@@ -63,33 +62,6 @@ public interface ITerminalControl {
 	Shell getShell();
 
 	/**
-	 * Set the encoding that the Terminal uses to decode bytes from the
-	 * Terminal-to-remote-Stream into Unicode Characters used in Java; or, to
-	 * encode Characters typed by the user into bytes sent over the wire to the
-	 * remote.
-	 *
-	 * By default, the local Platform Default Encoding is used. Also note that
-	 * the encoding must not be applied in case the terminal stream is processed
-	 * by some data transfer protocol which requires binary data.
-	 *
-	 * Validity of the encoding set here is not checked. Since some encodings do
-	 * not cover the entire range of Unicode characters, it can happen that a
-	 * particular Unicode String typed in by the user can not be encoded into a
-	 * byte Stream with the encoding specified. and UnsupportedEncodingException
-	 * will be thrown in this case at the time the String is about to be
-	 * processed.
-	 *
-	 * The concrete encoding to use can either be specified manually by a user,
-	 * by means of a dialog, or a connector can try to obtain it automatically
-	 * from the remote side e.g. by evaluating an environment variable such as
-	 * LANG on UNIX systems.
-	 *
-	 * @deprecated Use {@link #setCharset(Charset)} and do the error handling in the UI code.
-	 */
-	@Deprecated
-	void setEncoding(String encoding) throws UnsupportedEncodingException;
-
-	/**
 	 * Set the charset that the Terminal uses to decode bytes from the
 	 * Terminal-to-remote-Stream into Unicode Characters used in Java; or, to
 	 * encode Characters typed by the user into bytes sent over the wire to the
@@ -118,18 +90,6 @@ public interface ITerminalControl {
 	void setCharset(Charset charset);
 
 	/**
-	 * Return the current encoding. That's interesting when the previous
-	 * setEncoding() call failed and the fallback default encoding should be
-	 * queried, such that e.g. a combobox with encodings to choose can be
-	 * properly initialized.
-	 *
-	 * @return the current Encoding of the Terminal.
-	 * @deprecated Use {@link #getCharset()} and call {@link Charset#name()} on the result
-	 */
-	@Deprecated
-	String getEncoding();
-
-	/**
 	 * Return the current charset.
 	 *
 	 * @return the non-<code>null</code> current charset of the Terminal
@@ -152,12 +112,6 @@ public interface ITerminalControl {
 	 * to remote is in {@link ITerminalConnector#getTerminalToRemoteStream()}.
 	 */
 	OutputStream getRemoteToTerminalOutputStream();
-
-	/**
-	 * @deprecated call {@link #setTerminalTitle(String, TerminalTitleRequestor)} instead
-	 */
-	@Deprecated(forRemoval = true)
-	void setTerminalTitle(String title);
 
 	/**
 	 * Set the title of the terminal view.
