+++ terminal/tests/org.eclipse.tm.terminal.test/src/org/eclipse/terminal/internal/test/ui/VT100DataSource.java
@@ -25,7 +25,7 @@ import java.io.OutputStream;
 import java.io.Reader;
 import java.nio.charset.StandardCharsets;
 
-import org.eclipse.terminal.internal.control.ITerminalListener3.TerminalTitleRequestor;
+import org.eclipse.terminal.internal.control.ITerminalListener.TerminalTitleRequestor;
 import org.eclipse.terminal.internal.control.impl.ITerminalControlForText;
 import org.eclipse.terminal.internal.emulator.VT100Emulator;
 import org.eclipse.terminal.internal.provisional.api.ITerminalConnector;
